pyproject.toml
[project]
name = "sirius-it"
version = "1.0.0"
description = ""
authors = [
    {name = "n9dn9",email = "romanngurev07@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
